save files

+++ src/cmd/compile/ast/expression_check_identifier.go
@@ -21,7 +21,7 @@ func (e *Expression) checkIdentifierExpression(block *Block) (*Type, error) {
 		return result, nil
 	case magicIdentifierLine:
 		e.Type = ExpressionTypeInt
-		e.Data = int32(e.Pos.Line)
+		e.Data = int64(e.Pos.Line)
 		result, _ := e.checkSingleValueContextExpression(block)
 		return result, nil
 	case magicIdentifierTime:
